UaaTokenStore: Don't fall over when failing to expire oauth codes
MySQL does row-level locking, which means when we delete expired oauth
codes in parallel we can sometimes hit a harmless deadlock (where the
two delete updates execute across the rows in different order
simultaneously).  This is not an issue as the winning update will expire
tokens as appropriate, plus the tokens are checked for expiry on use.

This commit is cherry-picked from 0144d1310434
( https://github.com/cloudfoundry/uaa/pull/772 )